Refactor, removed unused usings, moved message class to this file, deleted messagesystem
