--- EchoServer/MessageSystem.cs
@@ -1,44 +1,0 @@
-﻿/*
-    OpenEcho is a program to automate basic tasks at home all while being handsfree.
-    Copyright (C) 2015 Gregory Morgan
-
-    This program is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    This program is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-using System;
-
-namespace EchoServer
-{
-    public class MessageSystem
-    {
-        public Message GetNextMessage() // called from server
-        {
-            // also mark as processing.
-            return SQL.GetNextMessage();
-        }
-    }
-
-    public class Message
-    {
-        public Guid clientID = Guid.Empty;
-        public Guid messageID = Guid.Empty;
-        public string textRequest = string.Empty;
-        public string textResponse = string.Empty;
-        public byte[] audioRequest;
-        public byte[] audioResponse = new byte[0];
-        public DateTime postTime = new DateTime();
-        public enum Status { queued, processing, delayed, ready, closed, error };
-        public Status status = Status.queued;
-    }
-}
