+++ EchoServer/Program.cs
@@ -16,85 +16,75 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+using Extensions;
+using PluginContracts;
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.IO;
-using Extensions;
-using System.Reflection;
-using System.Windows.Forms;
+using System.Linq;
 using System.Speech.Synthesis;
 using System.Threading;
-using PluginContracts;
-using System.Diagnostics;
-using System.Net;
-using System.Runtime.Serialization.Formatters.Binary;
-using System.Runtime.Serialization;
-using System.Runtime.Serialization.Json;
-using Newtonsoft.Json;
+using System.Threading.Tasks;
 
 namespace EchoServer
 {
     public class Program
     {
-        private QueryClassification qc = new QueryClassification();
-        private ResponseTime responseTime = new ResponseTime();
-        private MessageSystem messageSystem = new MessageSystem();
-
         public Program()
         {
-            Dictionary<string, IPlugin> _Plugins = LoadPlugins();
+            MessageSystem messageSystem = new MessageSystem();
+            ResponseTime responseTime = new ResponseTime();
+            QueryClassification qc = new QueryClassification();
+            Dictionary<string, IPlugin> _Plugins = LoadPlugins(qc);
 
             while (true)
             {
-                Message message = messageSystem.GetNextMessage();
+                Message message = SQL.GetNextMessage();
 
                 if (message == null || message.messageID == Guid.Empty)
                 {
                     Thread.Sleep(10);
                     continue;
                 }
-
-                if (string.IsNullOrEmpty(message.textRequest))
+                Task.Factory.StartNew(() =>
                 {
-                    message.textRequest = SpeechToText.Process(message.audioRequest);
-                }
-
-
-                KeyValuePair<string, string> query = qc.Classify(message.textRequest);
-
-                int responseTimeID = responseTime.Start(query.Key);
-                string delaymsg = responseTime.GetDelayMessage(query.Key);
-                if (!string.IsNullOrWhiteSpace(delaymsg))
-                {
-                    message.textResponse = delaymsg;
-                    message.audioResponse = GetAudio(delaymsg);
-                    message.status = Message.Status.delayed;
-                }
-
-                if (query.Key == "help")
-                {
-                    message.textResponse = query.Value;
-                    message.audioResponse = GetAudio(query.Value);
-                    message.status = Message.Status.ready;
-                }
-                else if (query.Key == "unknown")
-                {
-                    message.textResponse = query.Value;
-                    message.audioResponse = GetAudio(query.Value);
-                    message.status = Message.Status.ready;
-                }
-                else if (query.Key == "blank")
-                {
-                    message.status = Message.Status.error;
-                }
-                else
-                {
-                    if (_Plugins.ContainsKey(query.Key))
+                    if (string.IsNullOrEmpty(message.textRequest))
                     {
-                        string request = message.textRequest.Replace(query.Value, "").Trim();
+                        message.textRequest = SpeechToText.Process(message.audioRequest);
+                    }
+
+
+                    KeyValuePair<string, string> query = qc.Classify(message.textRequest);
+
+                    int responseTimeID = responseTime.Start(query.Key);
+                    string delaymsg = responseTime.GetDelayMessage(query.Key);
+                    if (!string.IsNullOrWhiteSpace(delaymsg))
+                    {
+                        message.textResponse = delaymsg;
+                        message.audioResponse = GetAudio(delaymsg);
+                        message.status = Message.Status.delayed;
+                    }
+
+                    if (query.Key == "help")
+                    {
+                        message.textResponse = query.Value;
+                        message.audioResponse = GetAudio(query.Value);
+                        message.status = Message.Status.ready;
+                    }
+                    else if (query.Key == "unknown")
+                    {
+                        message.textResponse = query.Value;
+                        message.audioResponse = GetAudio(query.Value);
+                        message.status = Message.Status.ready;
+                    }
+                    else if (query.Key == "blank")
+                    {
+                        message.status = Message.Status.error;
+                    }
+                    else if (_Plugins.ContainsKey(query.Key))
+                    {
+                        string request = message.textRequest.Replace(query.Value, "").CleanText();
+                        request = request.Length > 0 ? request : message.textRequest;
                         IPlugin plugin = _Plugins[query.Key];
                         string response = plugin.Go(request);
                         message.textResponse = response;
@@ -107,16 +97,16 @@ namespace EchoServer
                         message.audioResponse = GetAudio("A loaded plugin has failed to produce a valid key."); // very unlikely.
                         message.status = Message.Status.ready;
                     }
-                }
 
-                // post the message back to SQL.
-                message.audioRequest = new byte[0]; // remove request audio.
-                SQL.UpdateMessage(message);
-                responseTime.Stop(query.Key, responseTimeID);
+                    // post the message back to SQL.
+                    message.audioRequest = new byte[0]; // remove request audio.
+                    SQL.UpdateMessage(message);
+                    responseTime.Stop(query.Key, responseTimeID);
+                });
             }
         }
 
-        private Dictionary<string, IPlugin> LoadPlugins()
+        private Dictionary<string, IPlugin> LoadPlugins(QueryClassification qc)
         {
             Dictionary<string, IPlugin> _Plugins = new Dictionary<string, IPlugin>();
             string path = @"C:\Program Files\EchoServer\Plugins";
@@ -127,14 +117,14 @@ namespace EchoServer
             {
                 foreach (var item in plugins)
                 {
-                        _Plugins.Add(item.Name, item);
+                    _Plugins.Add(item.Name, item);
 
-                        //List<string> actions = _Plugins[item.Name].Actions;
+                    //List<string> actions = _Plugins[item.Name].Actions;
 
-                        foreach (string action in item.Actions)
-                        {
-                            qc.AddPhraseToAction(action, item.Name);
-                        }
+                    foreach (string action in item.Actions)
+                    {
+                        qc.AddPhraseToAction(action, item.Name);
+                    }
                 }
             }
             else
@@ -163,9 +153,22 @@ namespace EchoServer
                 t.Start();
                 t.Join();
             }
-           
+
 
             return buffer;
         }
+    }
+
+    public class Message
+    {
+        public Guid clientID = Guid.Empty;
+        public Guid messageID = Guid.Empty;
+        public string textRequest = string.Empty;
+        public string textResponse = string.Empty;
+        public byte[] audioRequest;
+        public byte[] audioResponse = new byte[0];
+        public DateTime postTime = new DateTime();
+        public enum Status { queued, processing, delayed, ready, closed, error };
+        public Status status = Status.queued;
     }
 }
